HF-110 [Refactor] Footer 컴포넌트로 변경 2023/05/05 - 21:32
