+++ hafit-fe/src/pages/JoinPage.jsx
@@ -10,11 +10,12 @@ import {
 } from "antd";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import axios from 'axios';
+import axios from "axios";
 
-import PhoneNumberInput from "../components/inputs/phoneNumberInput";
+import PhoneNumberInput from "../components/inputs/PhoneNumberInput";
+import Footer from "../components/Footer";
 
-import "../styles/joinPage.css";
+import "../styles/pages/joinPage.css";
 
 const { Option } = Select;
 const { Title } = Typography;
@@ -57,97 +58,100 @@ const JoinPage = () => {
   };
 
   return (
-    <div>
-      <Row justify="center" align="middle" style={{ height: "100vh" }} >
-        <Col span={12} align="middle">
-          <Title level={3}>회원가입</Title>
-          <Form form={form}
-          onFinish={onFinish}
-          onFinishFailed={onFinishFailed}
-          layout="vertical">
-            <Form.Item
-              label="이메일"
-              name="email"
-              rules={[
-                {
-                  required: true,
-                  message: "이메일을 입력해주세요",
-                },
-                {
-                  type: "email",
-                  message: "유효한 이메일을 입력해주세요",
-                },
-              ]}
+    <div className="top-container">
+      <div className="body-wrapper">
+        <Row justify="center" align="middle" style={{ height: "100vh" }}>
+          <Col span={12} align="middle">
+            <Title level={3}>회원가입</Title>
+            <Form
+              form={form}
+              onFinish={onFinish}
+              onFinishFailed={onFinishFailed}
+              layout="vertical"
             >
-              <Input placeholder="example@gmail.com" />
-            </Form.Item>
-            <Form.Item
-              label="이름"
-              name="name"
-              rules={[
-                {
-                  required: true,
-                  message: "이름을 입력해주세요",
-                },
-              ]}
-            >
-              <Input placeholder="홍길동" />
-            </Form.Item>
-            <Form.Item
-              label="비밀번호"
-              name="password"
-              rules={[
-                {
-                  required: true,
-                  message: "비밀번호를 입력해주세요",
-                },
-                {
-                  min: 8,
-                  message: "비밀번호는 최소 8자리 이상이어야 합니다",
-                },
-              ]}
-            >
-              <Input.Password placeholder="영문, 숫자, 특수문자 포함 8자 이상" />
-            </Form.Item>
-            <Form.Item
-              label="비밀번호 확인"
-              name="passwordConfirm"
-              rules={[
-                {
-                  required: true,
-                  message: "비밀번호 확인을 입력해주세요",
-                },
-                ({ getFieldValue }) => ({
-                  validator(_, value) {
-                    if (!value || getFieldValue("password") === value) {
-                      return Promise.resolve();
-                    }
-                    return Promise.reject(
-                      new Error("비밀번호가 일치하지 않습니다")
-                    );
+              <Form.Item
+                label="이메일"
+                name="email"
+                rules={[
+                  {
+                    required: true,
+                    message: "이메일을 입력해주세요",
                   },
-                }),
-              ]}
-            >
-              <Input.Password />
-            </Form.Item>
-            <Form.Item
-              label="통신사 선택"
-              name="carrier"
-              rules={[
-                {
-                  required: true,
-                  message: "통신사를 선택해주세요",
-                },
-              ]}
-            >
-              <Select placeholder="통신사 선택">
-                <Option value="SKT">SKT</Option>
-                <Option value="KT">KT</Option>
-                <Option value="LG">LG U+</Option>
-              </Select>
-            </Form.Item>
-            {/* <Form.Item
+                  {
+                    type: "email",
+                    message: "유효한 이메일을 입력해주세요",
+                  },
+                ]}
+              >
+                <Input placeholder="example@gmail.com" />
+              </Form.Item>
+              <Form.Item
+                label="이름"
+                name="name"
+                rules={[
+                  {
+                    required: true,
+                    message: "이름을 입력해주세요",
+                  },
+                ]}
+              >
+                <Input placeholder="홍길동" />
+              </Form.Item>
+              <Form.Item
+                label="비밀번호"
+                name="password"
+                rules={[
+                  {
+                    required: true,
+                    message: "비밀번호를 입력해주세요",
+                  },
+                  {
+                    min: 8,
+                    message: "비밀번호는 최소 8자리 이상이어야 합니다",
+                  },
+                ]}
+              >
+                <Input.Password placeholder="영문, 숫자, 특수문자 포함 8자 이상" />
+              </Form.Item>
+              <Form.Item
+                label="비밀번호 확인"
+                name="passwordConfirm"
+                rules={[
+                  {
+                    required: true,
+                    message: "비밀번호 확인을 입력해주세요",
+                  },
+                  ({ getFieldValue }) => ({
+                    validator(_, value) {
+                      if (!value || getFieldValue("password") === value) {
+                        return Promise.resolve();
+                      }
+                      return Promise.reject(
+                        new Error("비밀번호가 일치하지 않습니다")
+                      );
+                    },
+                  }),
+                ]}
+              >
+                <Input.Password />
+              </Form.Item>
+              <Form.Item
+                label="통신사 선택"
+                name="carrier"
+                rules={[
+                  {
+                    required: true,
+                    message: "통신사를 선택해주세요",
+                  },
+                ]}
+              >
+                <Select placeholder="통신사 선택">
+                  <Option value="SKT">SKT</Option>
+                  <Option value="KT">KT</Option>
+                  <Option value="LG">LG U+</Option>
+                </Select>
+              </Form.Item>
+              {/* <Form.Item
               label="전화번호"
               name="phoneNumber"
               rules={[
@@ -159,26 +163,28 @@ const JoinPage = () => {
             >
               <Input placeholder="010-1234-5678" />
             </Form.Item> */}
-            <PhoneNumberInput />
-            <Form.Item>
-              <Checkbox checked={agreed} onChange={handleAgreeChange}>
-                약관에 동의합니다
-              </Checkbox>
-            </Form.Item>
-            <Form.Item>
-              <Button 
-              type="primary" 
-              htmlType="submit" 
-              disabled={!agreed}
-              loading={loading}
-              block
-            >
-                가입하기
-              </Button>
-            </Form.Item>
-          </Form>
-        </Col>  
-      </Row>
+              <PhoneNumberInput />
+              <Form.Item>
+                <Checkbox checked={agreed} onChange={handleAgreeChange}>
+                  약관에 동의합니다
+                </Checkbox>
+              </Form.Item>
+              <Form.Item>
+                <Button
+                  type="primary"
+                  htmlType="submit"
+                  disabled={!agreed}
+                  loading={loading}
+                  block
+                >
+                  가입하기
+                </Button>
+              </Form.Item>
+            </Form>
+          </Col>
+        </Row>
+        <Footer />
+      </div>
     </div>
   );
 };
