+++ hafit-fe/src/styles/components/footer.css
@@ -1,0 +1,11 @@
+.footer {
+  text-align: center;
+  padding: 8px 0;
+  background-color: transparent;
+
+  width: 100%;
+  height: 48px; /* footer의 높이 */
+  position: absolute;
+  bottom: 0;
+  left: 0;
+}
