+++ hafit-fe/src/styles/components/navbar.css
@@ -3,11 +3,8 @@ body {
 }
 
 .navbar {
-  height: 70px;
-  min-height: 70px;
-  margin-left: 4rem;
-  margin-right: 4rem;
-  margin-top: 8px;
+  height: 100%;
+  max-height: 60px;
 }
 
 @media (max-width: 768px) {
@@ -24,6 +21,7 @@ body {
 
 .logo {
   margin-right: 16px;
+  margin-left: 4rem;
   width: 100px;
   height: 60px;
 }
@@ -31,6 +29,10 @@ body {
 .nav-menu {
   font-size: 16px; 
   font-weight: 500;
+}
+
+.menu-right {
+  margin-right: 4rem;
 }
 
 /* :where(.css-dev-only-do-not-override-htwhyh).ant-menu-light.ant-menu-horizontal >.ant-menu-item-selected::after, :where(.css-dev-only-do-not-override-htwhyh).ant-menu-light>.ant-menu.ant-menu-horizontal >.ant-menu-item-selected::after, :where(.css-dev-only-do-not-override-htwhyh).ant-menu-light.ant-menu-horizontal >.ant-menu-submenu-selected::after, :where(.css-dev-only-do-not-override-htwhyh).ant-menu-light>.ant-menu.ant-menu-horizontal >.ant-menu-submenu-selected::after {
