+++ hafit-fe/src/styles/pages/landingPage.css
@@ -4,31 +4,21 @@ html, body {
 
 /* START :: footer 하단 고정 */
 .top-container {
-    min-height: 96vh;
+    min-height: 86vh;
     position: relative;
 }
 
 .body-wrapper {
-    padding-bottom: 60px; /* footer의 높이 */
-}
-
-.footer {
-    text-align: center;
-    padding: 40px 0;
-    background-color: transparent;
-
-    width: 100%;
-    height: 60px; /* footer의 높이 */
-    position: absolute;  
-    bottom: 0;
-    left: 0;
+    padding-bottom: 48px; /* footer의 높이 */
+    margin-left: 4rem;
+    margin-right: 4rem;
 }
 /* END :: footer 하단 고정 */
 
 /* Header */
 .header {
   height: 80vh;
-  background-image: linear-gradient(to bottom, rgba(0, 0, 0, 0.85), rgba(0, 0, 0, 0.75), rgba(0, 0, 0, 0.55), rgba(255, 255, 255, 0.98)), url("../assets/img/isometric_gym_fitness_illust.png");
+  background-image: linear-gradient(to bottom, rgba(0, 0, 0, 0.85), rgba(0, 0, 0, 0.75), rgba(0, 0, 0, 0.55), rgba(255, 255, 255, 0.98)), url("../../assets/img/isometric_gym_fitness_illust.png");
   background-size: cover;
   display: flex;
   align-items: center;
