+++ hafit-fe/src/styles/pages/loginPage.css
@@ -99,20 +99,6 @@ label {
   padding-bottom: 48px; /* footer의 높이 */
   margin-left: 4rem;
   margin-right: 4rem;
-  align-items: center;
-  justify-content: center;
-}
-
-.footer {
-  text-align: center;
-  padding: 8px 0;
-  background-color: transparent;
-
-  width: 100%;
-  height: 48px; /* footer의 높이 */
-  position: absolute;
-  bottom: 0;
-  left: 0;
 }
 /* END :: footer 하단 고정 */
